Introducing an intermediate class
